Class Weapon is now abstract
Since enums limit what a weapon can and cannot do, the concept of "a weapon" should be abstract and only exist as a superclass of specific weapons, like swords, axes, knives, staves, bows and any others that might be added in the future.
These subclasses should be able to tell whoever tries to equip them whether they can do so or not, and each specific weapon has different users.

Signed-off-by: Tchy258 <tchy258@yahoo.cl>