Added more diversity and less idle time
Added all classes to the examples, only to illustrate.
Enemies were also added to the turn examples.
Added a turn count.
A new queue was made which stores all the weights of the weapons and enemies in ascending order, these values are used to determine how much time should actually pass in-between turns instead of a single Thread.sleep call.
An extra 100ms were added to each turn's wait time to have a wider margin of error.

Signed-off-by: Tchy258 <tchy258@yahoo.cl>