refactor: Proper double dispatch implementation
All weapons now override methods to equip themselves to a specific character class,
unless the weapon cannot be equipped to said character, in which case, the base
implementation throws an InvalidWeaponException.
Each of the PlayerCharacter implementations override the equip method to be able
to tell the weapon to be equipped what they are (a knight, a thief, etc).
And the weapon answers telling the character what they have to equip

Signed-off-by: Tchy258 <tchy258@yahoo.cl>